Add files via upload
The function names might be little bit different from website to avoid duplicate function names because same names were used for different problems on website. Very basic logical statements are used so the code might not be optimal, but yeah it works properly. Feel free to contact if code is not working in that website.